// api/src/utils/interfaces/common.ts
export interface CreateHospitalDto {
  name: string;
  province?: string;
  district?: string;
  sector?: string;
  cell?: string;
  village?: string;
  contact?: string;
  email?: string;
  chwSupervisor?: string;
  chwSupervisorContact?: string;
  totalChws?: number;
  activeChws?: number;
  catchmentArea?: string[];
}

export interface UpdateHospitalDto extends Partial<CreateHospitalDto> {}